more refactor into .rs files
