+++ src/database.rs
@@ -8,13 +8,7 @@ use serde::Deserialize;
 use serde_json;
 use smallvec::SmallVec;
 use std::collections::HashMap;
-use std::fs::File;
-use std::io::BufRead;
-use std::io::BufReader;
-use std::io::IoSliceMut;
 use std::io::Read;
-use std::io::Seek;
-use std::io::SeekFrom;
 use std::path::Path;
 use std::path::PathBuf;
 use std::sync::atomic::AtomicU32;
@@ -24,6 +18,11 @@ use zip;
 use zip::ZipArchive;
 
 use crate::blockhash::BlockIdHash;
+use crate::ziparchive::BlockLocation;
+use crate::ziparchive::MyCloneFileConfig;
+use crate::ziparchive::MyCloneFileReader;
+use crate::ziparchive::ZipArchiveWrapper;
+use crate::ziparchive::ZipLocation;
 
 #[derive(Deserialize)]
 #[allow(dead_code)] // Will use all these fields in the future
@@ -42,27 +41,6 @@ struct Manifest {
     pub(self) file_hash: String,
     #[serde(rename = "AppVersion")]
     pub(self) app_version: String,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq, PartialOrd, Ord)]
-/// Path to dblock.zip
-pub struct ZipLocation {
-    pub path_str: String,
-    pub path: PathBuf,
-}
-
-#[derive(Clone, Debug, PartialEq, PartialOrd, Eq, Ord)]
-pub struct BlockLocation {
-    /// Which file inside the zip
-    pub file_index: u32,
-
-    /// Which dblock.zip file
-    pub zip_path: Arc<ZipLocation>,
-}
-
-pub struct ZipArchiveWrapper {
-    pub zip_path: Arc<ZipLocation>,
-    archive: ZipArchive<MyCloneFileReader>,
 }
 
 pub struct HashToPath {
@@ -189,9 +167,10 @@ impl DFileDatabase {
                 )
                 .progress_chars("##-"),
         );
-        paths
-            .par_iter()
-            .try_for_each(|zip_path| self.import_from_zip(zip_path))?;
+        paths.par_iter().try_for_each(|zip_path| {
+            self.import_from_zip(zip_path)
+                .wrap_err_with(|| format!("import_from_zip: {:?}", zip_path))
+        })?;
 
         Ok(())
     }
@@ -205,8 +184,6 @@ impl DFileDatabase {
         });
         let zipbuf = MyCloneFileReader::new(config.clone())?;
         let zip = zip::ZipArchive::new(zipbuf)?;
-
-        let zip_len = zip.len();
 
         let arc_zippath = Arc::new(ZipLocation {
             path: zip_path.clone(),
@@ -236,77 +213,6 @@ impl DFileDatabase {
             }
         }
 
-        if false {
-            let mut hvec = Vec::new();
-            let (sender, receiver) = crossbeam_channel::bounded(zip_len + 1);
-
-            // make workers
-            for _t in 0..16 {
-                //println!("Make worker {}", t);
-
-                let receiver = receiver.clone(); // clone for this thread
-
-                //let zip_path = zip_path.clone();
-                let inner = self.inner.clone();
-                let mut zip = zip.clone();
-                let arc_zippath = arc_zippath.clone();
-                let handler = std::thread::spawn(move || {
-                    let mut progress_burst = 0;
-
-                    loop {
-                        let r = receiver.recv();
-                        match r {
-                            Ok(file_index) => {
-                                //let s = rng.gen_range(100..1000);
-
-                                //thread::sleep(Duration::from_millis(s));
-
-                                let hash_path =
-                                    zip.by_index(file_index).unwrap().name().to_string();
-                                let hash =
-                                    base64::decode_config(&hash_path, base64::URL_SAFE).unwrap();
-                                {
-                                    let mut inner = inner.lock().unwrap();
-
-                                    if let Some(hash2path) = &mut inner.hash2path {
-                                        hash2path.hash2path.insert(
-                                            hash.into(),
-                                            BlockLocation {
-                                                zip_path: arc_zippath.clone(),
-                                                file_index: file_index as u32,
-                                            },
-                                        );
-                                    }
-                                }
-
-                                progress_burst += 1;
-                                if progress_burst > 1000 {
-                                    //pb.inc(progress_burst);
-                                    println!(
-                                        "found hash {}/{} in {:?}",
-                                        file_index, zip_len, arc_zippath.path
-                                    );
-                                    progress_burst = 0;
-                                }
-                            }
-                            _ => break,
-                        }
-                    }
-                    //pb.inc(progress_burst);
-                });
-
-                hvec.push(handler);
-            }
-
-            for i in 0..zip_len {
-                sender.send(i).unwrap();
-            }
-            drop(sender);
-
-            for h in hvec {
-                h.join().unwrap();
-            }
-        }
         {
             use std::sync::atomic::Ordering;
             config.buf_capacity.store(32 * 1024, Ordering::Relaxed);
@@ -318,27 +224,7 @@ impl DFileDatabase {
             let path_str = arc_zippath.path_str.clone();
             inner.zip2ziparchive.insert(path_str, wrapper);
         }
-        // let paths: Vec<String> = (0..zip_len)
-        //     .into_par_iter()
-        //     .map_init(
-        //         || {
 
-        //             zip
-        //         },
-        //         |zip, i| ,
-        //     )
-        //     .collect();
-
-        // let bytes = zippath.as_bytes();
-        // for p in paths {
-
-        //     //conn.kv_store(hash, bytes).unwrap();
-
-        //     //println!("len: {}", hash.len());
-        //     //println!("zippath:{} hash:{}", zippath, p);
-
-        // }
-        //conn.commit().unwrap();
         Ok(())
     }
 
@@ -399,79 +285,5 @@ impl DFileDatabase {
 
     pub fn hash_size(&self) -> usize {
         32
-    }
-}
-pub struct MyCloneFileConfig {
-    pub path: PathBuf,
-    /// Changes after the files are indexed.
-    /// Bigger buf helps with large file reads.
-    /// Smaller buf does less redundant byte reads from disk when indexing.
-    pub buf_capacity: AtomicU32,
-}
-pub struct MyCloneFileReader {
-    pub config: Arc<MyCloneFileConfig>,
-    buf_reader: BufReader<File>,
-}
-
-impl Clone for MyCloneFileReader {
-    fn clone(&self) -> Self {
-        Self::new(self.config.clone()).unwrap()
-    }
-}
-
-impl MyCloneFileReader {
-    pub fn new(config: Arc<MyCloneFileConfig>) -> Result<Self> {
-        let target_file = File::open(&config.path)?;
-        let cap = config
-            .buf_capacity
-            .load(std::sync::atomic::Ordering::Relaxed);
-        let filebuf = BufReader::with_capacity(cap as usize, target_file);
-
-        Ok(Self {
-            config: config.clone(),
-            buf_reader: filebuf,
-        })
-    }
-}
-
-impl Read for MyCloneFileReader {
-    fn read(&mut self, buf: &mut [u8]) -> std::io::Result<usize> {
-        self.buf_reader.read(buf)
-    }
-
-    fn read_exact(&mut self, buf: &mut [u8]) -> std::io::Result<()> {
-        self.buf_reader.read_exact(buf)
-    }
-
-    fn read_vectored(&mut self, bufs: &mut [IoSliceMut<'_>]) -> std::io::Result<usize> {
-        self.buf_reader.read_vectored(bufs)
-    }
-
-    fn read_to_end(&mut self, buf: &mut Vec<u8>) -> std::io::Result<usize> {
-        self.buf_reader.read_to_end(buf)
-    }
-
-    fn read_to_string(&mut self, buf: &mut String) -> std::io::Result<usize> {
-        self.buf_reader.read_to_string(buf)
-    }
-}
-
-impl Seek for MyCloneFileReader {
-    fn seek(&mut self, pos: SeekFrom) -> std::io::Result<u64> {
-        self.buf_reader.seek(pos)
-    }
-
-    fn stream_position(&mut self) -> std::io::Result<u64> {
-        self.buf_reader.stream_position()
-    }
-}
-
-impl BufRead for MyCloneFileReader {
-    fn fill_buf(&mut self) -> std::io::Result<&[u8]> {
-        self.buf_reader.fill_buf()
-    }
-
-    fn consume(&mut self, amt: usize) {
-        self.buf_reader.consume(amt)
     }
 }
